requirements.txt
Django>=5.1.2,<5.2
django-extensions>=3.2.3,<3.3
python-decouple==3.8
requests>=2.32.3,<2.33
django-filter==24.3
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.3.0,<5.4
psycopg2-binary>=2.9.9,<2.10 #in production compile all the libs necessary and use psycopg2 lib
Markdown==3.7

# For Development
django-seed==0.3.1 # python manage.py seed api --number=15
Faker==37.4.0 # para uso na criação de dados falsos controlados
coverage==7.6.3 # coverage run manage.py test .......... coverage report -m